v4/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
Authlib==1.3.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.1
distro==1.9.0
Flask==3.1.0
frozenlist==1.5.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-health-checking==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.19
langchain-core==0.3.36
langchain-openai==0.3.6
langchain-text-splitters==0.3.6
langchain-weaviate==0.0.4
langsmith==0.3.8
MarkupSafe==3.0.2
multidict==6.1.0
numpy==1.26.4
openai==1.63.2
orjson==3.10.15
packaging==24.2
propcache==0.2.1
protobuf==5.29.3
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==75.8.0
simsimd==6.2.1
sniffio==1.3.1
SQLAlchemy==2.0.38
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
validators==0.34.0
weaviate-client==4.11.0
Werkzeug==3.1.3
yarl==1.18.3
zstandard==0.23.0
